Fix package problems
Thanks IDEA, very nice refactor.

+++ src/main/java/com/github/frontear/mixins/impl/MixinRender.java
@@ -8,8 +8,6 @@ import net.minecraft.client.renderer.*;
 import net.minecraft.client.renderer.entity.*;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.entity.*;
-import org.frontear.infinity.Infinity;
-import org.frontear.infinity.modules.impl.HealthTag;
 import org.lwjgl.opengl.GL11;
 import org.spongepowered.asm.mixin.*;
 
@@ -63,16 +61,16 @@ public abstract class MixinRender {
             int j = fontrenderer.getStringWidth(str) / 2;
             GlStateManager.disableTexture2D();
             worldrenderer.begin(7, DefaultVertexFormats.POSITION_COLOR);
-            worldrenderer.pos((double) (-j - 1), (double) (-1 + i), 0.0D)
+            worldrenderer.pos(-j - 1, -1 + i, 0.0D)
                 .color(0.0F, 0.0F, 0.0F, 0.25F)
                 .endVertex();
-            worldrenderer.pos((double) (-j - 1), (double) (8 + i), 0.0D)
+            worldrenderer.pos(-j - 1, 8 + i, 0.0D)
                 .color(0.0F, 0.0F, 0.0F, 0.25F)
                 .endVertex();
-            worldrenderer.pos((double) (j + 1), (double) (8 + i), 0.0D)
+            worldrenderer.pos(j + 1, 8 + i, 0.0D)
                 .color(0.0F, 0.0F, 0.0F, 0.25F)
                 .endVertex();
-            worldrenderer.pos((double) (j + 1), (double) (-1 + i), 0.0D)
+            worldrenderer.pos(j + 1, -1 + i, 0.0D)
                 .color(0.0F, 0.0F, 0.0F, 0.25F)
                 .endVertex();
             tessellator.draw();
